+++ docs/changes.rst
@@ -40,6 +40,8 @@ Released: not yet
 
 **Enhancements:**
 
+* Properties in JSON output are now always sorted by property name. (issue #267)
+
 * Added support for the "console" command group, with the following commands:
 
   - get-audit-log     - Get the audit log of the targeted HMC.
@@ -47,6 +49,7 @@ Released: not yet
   - show              - Show properties of the console of the targeted HMC.
 
   Issue #277
+
 
 **Cleanup:**